+++ lib/src/apollovm_extension.dart
@@ -20,6 +20,27 @@ extension FutureOrExtension<T> on FutureOr<T> {
       return mapper(this as T);
     }
   }
+
+  FutureOr<R> resolveWith<R>(FutureOr<R> Function() mapper) {
+    if (this is Future) {
+      var future = this as Future<T>;
+      return future.then((r) => mapper());
+    } else {
+      return mapper();
+    }
+  }
+
+  FutureOr<void> onResolve<R>(void Function(T r) callback) {
+    if (this is Future) {
+      var future = this as Future<T>;
+      return future.then((r) => callback(r));
+    } else {
+      callback(this as T);
+    }
+  }
+
+  Future<T> get asFuture =>
+      this is Future ? this as Future<T> : Future<T>.value(this);
 }
 
 extension ListFutureOrExtension<T> on List<FutureOr<T>> {