+++ lib/src/ast/apollovm_ast_annotation.dart
@@ -6,6 +6,20 @@ class ASTAnnotation implements ASTNode {
   Map<String, ASTAnnotationParameter>? parameters;
 
   ASTAnnotation(this.name, [this.parameters]);
+
+  ASTNode? _parentNode;
+
+  @override
+  ASTNode? get parentNode => _parentNode;
+
+  @override
+  void resolveNode(ASTNode? parentNode) {
+    _parentNode = parentNode;
+  }
+
+  @override
+  ASTNode? getNodeIdentifier(String name) =>
+      parentNode?.getNodeIdentifier(name);
 }
 
 class ASTAnnotationParameter implements ASTNode {
@@ -17,4 +31,18 @@ class ASTAnnotationParameter implements ASTNode {
 
   ASTAnnotationParameter(this.name, this.value,
       [this.defaultParameter = false]);
+
+  ASTNode? _parentNode;
+
+  @override
+  ASTNode? get parentNode => _parentNode;
+
+  @override
+  void resolveNode(ASTNode? parentNode) {
+    _parentNode = parentNode;
+  }
+
+  @override
+  ASTNode? getNodeIdentifier(String name) =>
+      parentNode?.getNodeIdentifier(name);
 }