+++ lib/src/core/apollovm_core_base.dart
@@ -1,10 +1,14 @@
 import 'package:apollovm/apollovm.dart';
+import 'package:swiss_knife/swiss_knife.dart';
 
 class ApolloVMCore {
   static ASTClass<V>? getClass<V>(String className) {
     switch (className) {
       case 'String':
         return CoreClassString.INSTANCE as ASTClass<V>;
+      case 'int':
+      case 'Integer':
+        return CoreClassInt.INSTANCE as ASTClass<V>;
       default:
         return null;
     }
@@ -12,7 +16,9 @@ class ApolloVMCore {
 }
 
 abstract class CoreClassPrimitive<T> extends ASTClassPrimitive<T> {
-  CoreClassPrimitive(ASTTypePrimitive<T> type) : super(type) {
+  final String coreName;
+
+  CoreClassPrimitive(ASTTypePrimitive<T> type, this.coreName) : super(type) {
     type.setClass(this);
   }
 
@@ -73,7 +79,7 @@ class CoreClassString extends CoreClassPrimitive<String> {
 
   late final ASTExternalFunction _function_valueOf;
 
-  CoreClassString._() : super(ASTTypeString.INSTANCE) {
+  CoreClassString._() : super(ASTTypeString.INSTANCE, 'String') {
     _function_contains = _externalClassFunctionArgs1(
       'contains',
       ASTTypeBool.INSTANCE,
@@ -116,6 +122,45 @@ class CoreClassString extends CoreClassPrimitive<String> {
         return _function_valueOf;
     }
     throw StateError(
-        "Can't find core function: $fName( $parametersSignature )");
+        "Can't find core function: $coreName.$fName( $parametersSignature )");
+  }
+}
+
+class CoreClassInt extends CoreClassPrimitive<int> {
+  static final CoreClassInt INSTANCE = CoreClassInt._();
+
+  late final ASTExternalFunction _function_valueOf;
+
+  late final ASTExternalFunction _function_parseInt;
+
+  CoreClassInt._() : super(ASTTypeInt.INSTANCE, 'int') {
+    _function_parseInt = _externalStaticFunctionArgs1(
+      'parseInt',
+      ASTTypeInt.INSTANCE,
+      ASTFunctionParameterDeclaration(ASTTypeString.INSTANCE, 's', 0, false),
+      (dynamic p1) => parseInt(p1),
+    );
+
+    _function_valueOf = _externalStaticFunctionArgs1(
+      'valueOf',
+      ASTTypeString.INSTANCE,
+      ASTFunctionParameterDeclaration(ASTTypeDynamic.INSTANCE, 'obj', 0, false),
+      (dynamic o) => '$o',
+    );
+  }
+
+  @override
+  ASTFunctionDeclaration? getFunction(
+      String fName, ASTFunctionSignature parametersSignature, VMContext context,
+      {bool caseInsensitive = false}) {
+    switch (fName) {
+      case 'parseInt':
+      case 'parse':
+        return _function_parseInt;
+      case 'valueOf':
+        return _function_valueOf;
+    }
+    throw StateError(
+        "Can't find core function: $coreName.$fName( $parametersSignature )");
   }
 }